Add <p> Element Around Toast Content
Allow env value with new line characters to render appropriately.

Note however that the current tests will fail if your local env's value for
the message rendered on failure to create new chat contains new line characters (\n)!

This is because the test uses @react-testing-library.screen.getAllByText() to find the content
of the toast on the page.